backend/requirements.txt
asgiref==3.10.0
Django==5.2.8
django-split-settings==1.3.2
django-cors-headers==4.3.1
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.3
